Switch to core20 since this provides a newer, less noisy fontconfig
On Ubuntu 20.04 (which is most popular with emacs snap users) and the older
core18 base there is a lot of fontconfig related errors / noise at start
up (issue #10). Switching to core20 base gives no fontconfig errors on
Ubuntu 20.04 and only 1 on Ubuntu 18.04 so this is a reasonable solution
eliminate or reduce the number of warnings at startup for the majority of
users.