Unset various SNAP environment variables on startup
Firefox (and other applications) may change their behaviour when running as
a snap - as such, if we have these SNAP environment variables set in our
environment and then we go and execute say Firefox, it will assume it is
running as a snap and may then misbehave. This causes issues like Firefox
using the wrong profile as seen in
https://bugs.launchpad.net/snapd/+bug/1835024 and more specifically in
https://github.com/alexmurray/emacs-snap/issues/36. If we unset the
variables then we avoid any confusion in these other applications, plus
Emacs itself doesn't do anything with these anyway so it is safe to unset
them too.

Fixes #36.

Signed-off-by: Alex Murray <murray.alex@gmail.com>