Untested fixes to try and get tree-sitter working
Since emacs compiles and loads tree sitter parts as native modules, these need
to get compiled with the same compiler and libc etc as the emacs snap itself. So
ensure we include both gcc-10 and g++-10 in the snap and override treesit.el to
use these paths instead of trying to find gcc/g++ on the host system.

Should hopefully fix issue #55

Signed-off-by: Alex Murray <murray.alex@gmail.com>