Use --with-native-compilation=aot
It seems this got changed before emacs 29 was released so make sure we specify
the correct way to invoke this to ensure we pre-compile as much elisp as
possible. Also keep NATIVE_FULL_AOT for now as well as this is still referenced
in the upstream lisp/Makefile.in so it may still be required.

As seen in
https://www.reddit.com/r/emacs/comments/15g0gxt/emacs_291_compiling_some_packages_on_every_startup/

Should hopefully fix issue #64.

Signed-off-by: Alex Murray <murray.alex@gmail.com>