Fix linker script to search for libc.so.6 in multiple paths
By default the script was using the path for Ubuntu but this fails on other
distros like Rocky Linux where this does not exist. So instead use a relative
path for libc.so.6 and add both the Ubuntu and the Rocky path to the search path
for the linker so it can find either. Fixes issues #96 and #100.

Signed-off-by: Alex Murray <murray.alex@gmail.com>