setup-env: respect host systems gsettings as much as possible
Instead of just setting GSETTINGS_SCHEMA_DIR to to the one shipped with the
snap, try and first look to see if the one from the host has the required
elements, and if not then duplicate it into ~/snap/emacs/common/.cache/schemas
by symlinking to each file from the host, but then for each file which does not
have the required elements, symlink to the one from the snap. Then finally
compile them all via glib-compile-schemas and set GSETTINGS_SCHEMA_DIR to point
to the frankenschema dir ~/snap/emacs/common/.cache/schemas.

As such, we then use as many of the hosts schema overrides etc and so this
ensures we show maximize/minimize titlebar buttons on Ubuntu etc to fix
issue #104.

Signed-off-by: Alex Murray <murray.alex@gmail.com>