Build emacs-gtk before emacs-wayland
It is expected that the majority of the users of the snap are on a wayland
session (since we have 43% on 24.04, 27% on 22.04 and another 5% on the later
interim releases - so at least 75% of users are running an OS that defaults to
the wayland session).

It would appear that the issue in #106 occurs for whichever emacs build is done
first - so by building wayland second we can remove this issue for the majority
of users. It is unclear however why this does not occur on the snapshot branch
using emacs 31 etc but that is an issue for another day.

Signed-off-by: Alex Murray <murray.alex@gmail.com>